収支金額・貯金金額・貯金目標・月次取引金額をBigintに変更

+++ MoneyLogAPI/src/main/java/com/example/domain/CategoryList.java
@@ -1,45 +1,46 @@
 package com.example.domain;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public class CategoryList {
 
-	private String categoryName;
-	private Integer categoryTotalAmount;
-	private List<Transaction> subCategoryList;
+    private String categoryName;
+    private BigInteger categoryTotalAmount;
+    private List<Transaction> subCategoryList;
 
-	public String getCategoryName() {
-		return categoryName;
-	}
+    public String getCategoryName() {
+        return categoryName;
+    }
 
-	public void setCategoryName(String categoryName) {
-		this.categoryName = categoryName;
-	}
+    public void setCategoryName(String categoryName) {
+        this.categoryName = categoryName;
+    }
 
-	public Integer getCategoryTotalAmount() {
-		return categoryTotalAmount;
-	}
+    public BigInteger getCategoryTotalAmount() {
+        return categoryTotalAmount;
+    }
 
-	public void setCategoryTotalAmount(Integer categoryTotalAmount) {
-		this.categoryTotalAmount = categoryTotalAmount;
-	}
+    public void setCategoryTotalAmount(BigInteger categoryTotalAmount) {
+        this.categoryTotalAmount = categoryTotalAmount;
+    }
 
-	public List<Transaction> getSubCategoryList() {
-		return subCategoryList;
-	}
+    public List<Transaction> getSubCategoryList() {
+        return subCategoryList;
+    }
 
-	public void setSubCategoryList(List<Transaction> subCategoryList) {
-		this.subCategoryList = subCategoryList;
-	}
+    public void setSubCategoryList(List<Transaction> subCategoryList) {
+        this.subCategoryList = subCategoryList;
+    }
 
-	public CategoryList() {
-		super();
-	}
+    public CategoryList() {
+        super();
+    }
 
-	public CategoryList(String categoryName, Integer categoryTotalAmount, List<Transaction> subCategoryList) {
-		super();
-		this.categoryName = categoryName;
-		this.categoryTotalAmount = categoryTotalAmount;
-		this.subCategoryList = subCategoryList;
-	}
+    public CategoryList(String categoryName, BigInteger categoryTotalAmount, List<Transaction> subCategoryList) {
+        super();
+        this.categoryName = categoryName;
+        this.categoryTotalAmount = categoryTotalAmount;
+        this.subCategoryList = subCategoryList;
+    }
 }
