+++ MoneyLogAPI/src/main/java/com/example/response/GetMonthlyFixedSpendingResponse.java
@@ -1,28 +1,29 @@
 package com.example.response;
 
-import java.util.List;
-
 import com.example.domain.MonthlyFixedList;
+
+import java.math.BigInteger;
+import java.util.List;
 
 public class GetMonthlyFixedSpendingResponse extends response {
 
-	private Integer disposableIncome;
-	private List<MonthlyFixedList> monthlyFixedList;
+    private BigInteger disposableIncome;
+    private List<MonthlyFixedList> monthlyFixedList;
 
-	public Integer getDisposableIncome() {
-		return disposableIncome;
-	}
+    public BigInteger getDisposableIncome() {
+        return disposableIncome;
+    }
 
-	public void setDisposableIncome(Integer disposableIncome) {
-		this.disposableIncome = disposableIncome;
-	}
+    public void setDisposableIncome(BigInteger disposableIncome) {
+        this.disposableIncome = disposableIncome;
+    }
 
-	public List<MonthlyFixedList> getMonthlyFixedList() {
-		return monthlyFixedList;
-	}
+    public List<MonthlyFixedList> getMonthlyFixedList() {
+        return monthlyFixedList;
+    }
 
-	public void setMonthlyFixedList(List<MonthlyFixedList> monthlyFixedList) {
-		this.monthlyFixedList = monthlyFixedList;
-	}
+    public void setMonthlyFixedList(List<MonthlyFixedList> monthlyFixedList) {
+        this.monthlyFixedList = monthlyFixedList;
+    }
 
 }
