+++ MoneyLogAPI/src/main/java/com/example/response/GetTotalSavingResponse.java
@@ -1,34 +1,35 @@
 package com.example.response;
 
-import java.util.List;
-
 import com.example.domain.MonthlySavingData;
+
+import java.math.BigInteger;
+import java.util.List;
 
 public class GetTotalSavingResponse extends response {
 
-	private Integer totalSavingAmount;
-	private List<MonthlySavingData> savingDataList;
+    private BigInteger totalSavingAmount;
+    private List<MonthlySavingData> savingDataList;
 
-	public Integer getTotalSavingAmount() {
-		return totalSavingAmount;
-	}
+    public BigInteger getTotalSavingAmount() {
+        return totalSavingAmount;
+    }
 
-	public void setTotalSavingAmount(Integer totalSavingAmount) {
-		this.totalSavingAmount = totalSavingAmount;
-	}
+    public void setTotalSavingAmount(BigInteger totalSavingAmount) {
+        this.totalSavingAmount = totalSavingAmount;
+    }
 
-	public List<MonthlySavingData> getSavingDataList() {
-		return savingDataList;
-	}
+    public List<MonthlySavingData> getSavingDataList() {
+        return savingDataList;
+    }
 
-	public void setSavingDataList(List<MonthlySavingData> savingDataList) {
-		this.savingDataList = savingDataList;
-	}
+    public void setSavingDataList(List<MonthlySavingData> savingDataList) {
+        this.savingDataList = savingDataList;
+    }
 
-	@Override
-	public String toString() {
-		return "GetTotalSavingResponse [totalSavingAmount=" + totalSavingAmount + ", savingDataList=" + savingDataList
-				+ "]";
-	}
+    @Override
+    public String toString() {
+        return "GetTotalSavingResponse [totalSavingAmount=" + totalSavingAmount + ", savingDataList=" + savingDataList
+                + "]";
+    }
 
 }
